refactor(agent): group pi frontend chat helpers
