+++ packages/agent/src/front/ChatPanel.tsx
@@ -33,8 +33,8 @@ import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { MentionPicker, detectMention, type MentionState } from './primitives/mention-picker'
 import { SlashCommandPicker } from './primitives/slash-command-picker'
 import { useAgentChat } from './hooks/useAgentChat'
-import { usePiChatProjection } from './piChatProjection'
-import { usePiNativeFollowUpQueue } from './piNativeFollowUpQueue'
+import { usePiChatProjection } from './pi/piChatProjection'
+import { usePiNativeFollowUpQueue } from './pi/piNativeFollowUpQueue'
 import { PI_AGENT_RUNTIME_CAPABILITIES } from '../shared/capabilities'
 import { DebugDrawer } from './DebugDrawer'
 import { builtinCommands } from './slashCommands/builtins'
