+++ packages/agent/src/front/pi/piChatProjection.ts
